move InterpreterTestSet to its own module
